requirements.txt
gunicorn==20.1.0
Flask==2.2.3
flask-httpauth==4.7.0
Werkzeug==2.2.3
Jinja2==3.1.2
MarkupSafe==2.1.2
itsdangerous==2.1.2
click==8.1.3
python-dotenv==0.10.1
